+++ app/controllers/dashboard/youtube/channels/application_controller.rb
@@ -1,8 +1,0 @@
-class Dashboard::Youtube::Channels::ApplicationController < Dashboard::ApplicationController
-  before_action :set_youtube_channel
-
-  private
-    def set_youtube_channel
-      @youtube_channel = YoutubeChannel.find(params.expect(:channel_id))
-    end
-end