+++ app/controllers/dashboard/youtube_channels_controller.rb
@@ -1,4 +1,4 @@
-class Dashboard::Youtube::ChannelsController < Dashboard::ApplicationController
+class Dashboard::YoutubeChannelsController < Dashboard::ApplicationController
   before_action :set_youtube_channel, only: %i[ show edit update destroy generate_youtube_short ]
 
   def index
@@ -34,8 +34,8 @@ class Dashboard::Youtube::ChannelsController < Dashboard::ApplicationController
   end
 
   def destroy
-    @youtube_channel.destroy!
-    redirect_to youtube_channels_path, notice: "Youtube channel was successfully destroyed.", status: :see_other
+    # @youtube_channel.destroy!
+    redirect_to dashboard_youtube_channels_path, notice: "Youtube channel was successfully destroyed.", status: :see_other
   end
 
   def generate_youtube_short