+++ app/models/youtube_short.rb
@@ -1,14 +1,14 @@
 class YoutubeShort < ApplicationRecord
   belongs_to :youtube_channel
-  has_many_attached :video_files
+  has_one_attached :audio
+  has_one_attached :srt
+  has_one_attached :video
 
-  validates :subject, presence: true
-  validates :max_duration, numericality: { only_integer: true, greater_than: 0 }
-  validates :simultaneous_videos, numericality: { only_integer: true, greater_than: 0 }
+  validates :status, presence: true
 
-  broadcasts_to ->(youtube_short) { [ youtube_short.channel, :youtube_shorts ] }, inserts_by: :prepend, target: "youtube_shorts", partial: "dashboard/youtube/channels/shorts/short"
+  broadcasts_to ->(youtube_short) { [ youtube_short.youtube_channel, :youtube_shorts ] }, inserts_by: :prepend, target: "youtube_shorts", partial: "dashboard/youtube_channels/youtube_shorts/youtube_short"
 
   def to_partial_path
-    "dashboard/youtube/channels/shorts/short"
+    "dashboard/youtube_channels/youtube_shorts/youtube_short"
   end
 end