+++ app/services/youtube/shorts/creation_service.rb
@@ -1,8 +1,0 @@
-class Youtube::Shorts::CreationService
-  def initialize(youtube_channel)
-    @youtube_channel = youtube_channel
-  end
-
-  def call
-  end
-end